Moved timetable render funcs to Helper
These functions have been now generalized, meaning the render target can now be specified instead of being hardcoded. This is preparation for the Builder dialog, so that the same function can be used to render to both the main timetable display as well as the Builder dialog timetable preview.
